fix: make the Linux platform code compile
With the manifest fixed, Linux reached simon's own source for the first time and
found 73 compile errors across 30 files. None of this code had ever been built:
the manifest bug meant cargo never got far enough to try, and CI's ubuntu and macos
jobs had been failing on the missing crates long before reaching any of it.

The substantive ones:

- SimonError::IoError does not exist (23 sites, cpufreq.rs and fan_control.rs).
  The enum has Io(io::Error), which cannot carry the context string these call
  sites build, so they now use System(String). Adding a variant would have been a
  breaking change to a public enum for no gain.
- core/engine.rs called path_exists and read_file_u32 unqualified and never
  imported them.
- chrono was declared without its serde feature, so eight derives on DateTime<Utc>
  had no impl. Windows got it transitively; nothing guaranteed that.
- motherboard/linux.rs disagreed with the traits it implements on six fields and
  three variants. While mapping DisplayOutputType, note the eDP arm sat after the
  DP arm and "eDP-1" contains "DP" — every internal panel was reported as an
  external DisplayPort, and the eDP arm was unreachable. eDP is now tested first
  and maps to Internal, which is what the enum calls that.
- hwmon/smart.rs called glob::glob; glob is not a dependency and never was. The
  pattern is one directory level with a fixed prefix, so read_dir does it without
  taking a dependency for a single call.
- services.rs borrowed self mutably and immutably at once; numa/mod.rs referenced a
  binding one line after it left scope; process_monitor.rs divided f64 by u64.
- ai_workload.rs and sandbox.rs named parameters _pid and _info, then used pid and
  info in their Linux branches.

Verified with `cargo check --target x86_64-unknown-linux-gnu`, which needs only a
rust-std component and gave a local loop instead of guessing through CI. Windows
clippy and the full suite still pass; three bindings needed #[allow(unused_mut)] or
a cfg'd import, since they are only written on Linux.

+++ src/process_monitor.rs
@@ -1492,7 +1492,7 @@ mod linux {
                     .and_then(|s| s.parse::<u64>().ok())
             });
         let process_start_time =
-            boot_time_secs.map(|btime| btime + (starttime as f64 / clk_tck) as u64);
+            boot_time_secs.map(|btime| btime + (starttime as f64 / clk_tck as f64) as u64);
 
         // Read /proc/[pid]/statm for memory
         let statm_path = format!("{}/statm", proc_path);
